requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
certifi==2018.4.16
enum34==1.1.6
lazy-object-proxy==1.3.1
paho-mqtt==1.3.1
pycparser==2.18
pycrypto==2.6.1
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
